backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.1
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.26.2
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
ultralytics==8.0.227
python-dotenv==1.0.0
google-generativeai==0.7.2
trino==0.329.0
